requirements.txt
# Cleaned-up dependencies for Astro with dbt + Snowflake
apache-airflow-providers-snowflake